resolver: extract builders, catalog, and package-ext helpers
Further split lib.rs (2545 -> 2091 lines, -454) into three more modules:

  builder.rs      261  Resolver::new + with_stream + 20 with_* setters
  catalog.rs      115  resolve_catalog_spec + materialize_catalog_picks
  package_ext.rs  117  pick_override_spec, apply_package_extensions,
                       package_selector_matches, is_deprecation_allowed

Remaining bulk in lib.rs is the 1800-line resolve_workspace method.
Further carving its inlined branches would trade 50-line chunks for
10-arg helper signatures — tight state coupling means the payoff is
marginal, so stopping here.
